+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2016-01-07
+### Added
+- `permission().hasPrivilege()` method
+- `permission().hasPrivileges()` alias
+
 ## [1.0.0] - 2016-01-07
 ### Fixed
 - Fix parameters being validated in the wrong order